python/webhooks/requirements.txt
bottle==0.12.16
PyYAML==5.1
